added card woosh and saved readings buttoon
I added a woosh and deWoosh animation to the style sheet. I then made two functions, woosh and dewoosh in the script. dewoosh() is called to woosh the cards on the screen. dewoosh() is called to woosh cards that aren't selected to the side when prediction is being made. I added a saved readings button because code for it existed on the javascript file but the button didn't exist in the html file, which was causing errors